feat(account-pages): moved Menu component into separate file
