+++ packages/falcon-ui/src/components/Navbar.tsx
@@ -23,29 +23,25 @@ export const Navbar = themed({
   }
 });
 
-type NavbarContextType = {
-  open?: boolean;
-};
+const MenuItemContext = React.createContext<{ open?: boolean }>({});
 
-const NavbarItemContext = React.createContext<NavbarContextType>({});
-
-const NavbarItemInnerDOM: React.SFC<any> = props => (
+const MenuItemInnerDOM: React.SFC<any> = props => (
   <Value initial={false}>
     {({ set, value }) => (
-      <NavbarItemContext.Provider value={{ open: value }}>
+      <MenuItemContext.Provider value={{ open: value }}>
         <ListItem
           {...props}
           onMouseEnter={() => set(true)}
           onMouseLeave={() => set(false)}
           onClick={() => set(false)}
         />
-      </NavbarItemContext.Provider>
+      </MenuItemContext.Provider>
     )}
   </Value>
 );
 
 export const NavbarItem = themed({
-  tag: NavbarItemInnerDOM,
+  tag: MenuItemInnerDOM,
 
   defaultTheme: {
     navbarItem: {
@@ -68,11 +64,10 @@ export const NavbarItem = themed({
 });
 
 const NavbarItemMenuInnerDOM: React.SFC<any> = props => (
-  <NavbarItemContext.Consumer>
+  <MenuItemContext.Consumer>
     {({ open }) => <Box {...props} display={open ? 'block' : 'none'} />}
-  </NavbarItemContext.Consumer>
+  </MenuItemContext.Consumer>
 );
-
 export const NavbarItemMenu = themed({
   tag: NavbarItemMenuInnerDOM,
 
@@ -92,54 +87,6 @@ export const NavbarItemMenu = themed({
         left: 0,
         right: 0
       }
-    }
-  }
-});
-
-export const Menu = themed({
-  tag: 'ul',
-
-  defaultTheme: {
-    menu: {
-      p: 'none',
-      m: 'none',
-      bg: 'primary',
-      color: 'primaryText',
-      css: {
-        display: 'flex',
-        flexDirection: 'column',
-        listStyle: 'none',
-        position: 'relative'
-      }
-    }
-  }
-});
-
-export const MenuItem = themed({
-  tag: NavbarItemInnerDOM,
-
-  defaultTheme: {
-    menuItem: {
-      fontSize: 'sm',
-      fontWeight: 'bold',
-      display: 'flex',
-      color: 'primaryText',
-
-      css: ({ theme }) => ({
-        cursor: 'pointer',
-        userSelect: 'none',
-        listStyle: 'none',
-
-        ':not(:last-child)': {
-          borderBottom: theme.borders.regular,
-          borderColor: theme.colors.secondaryDark
-        },
-
-        ':hover': {
-          background: theme.colors.primaryLight,
-          borderColor: theme.colors.secondaryLight
-        }
-      })
     }
   }
 });
