+++ packages/falcon-ui/src/index.ts
@@ -30,6 +30,7 @@ export * from './components/NumberInput';
 export * from './components/Select';
 export * from './components/Breadcrumbs';
 export * from './components/Details';
+export * from './components/Menu';
 export * from './components/Navbar';
 export * from './components/Dropdown';
 export * from './components/ThemeProvider';
